Created new version of gpio.h
This branch is to showcase several ideas that I think would be useful in
RASWare and may not nescessarily need to be included.

1. I've added definitions for a tPin type that represents all possible
pins. This allows for passing pins as arguments into functions as single
arguments as shown in the new gpio.h file. This is how I think the pins
should be handled as opposed to the complex macros we currently have.

2. I've added GetPin and SetPin functions for setting and retrieving
individual pins without prior setup. This has been avoided in previous
RASWare versions due to the existance of GPIORead and GPIOWrite but I
think having versions that use the new pin arguments and save two lines
can be useful and the fact that if works on indivisual pins welcoming to
unfamiliar raslets.

3. I've also modified the interrupt handlers to use individual pins. This
means you can only register a single interrupt per pin, however the
ability to register on the specific pin is worth the trade off in my
opinion.

I have not tested the new implementations and it does not yet compile due
to changed file names. However it should be functional as is.