Added pwm phase and frequency manipulation per Frank's request.
He pointed out that it was nescessary for the motor code to modify pwm's
phase and period. So as a solution they are now variables in pwm and can
be changed with SetPWMFrequency and SetPWMPhase. If the added runtime
usage proves problematic, seperating pwm into both motor and servo modules
will be nescessary as Frank originally suggested.