Removed unescessary TimeS functions and added US macro
TimeS functions are pretty useless as the float accepting functions are in
seconds. TimeUS functions should still be kept as they allow for no
overhead when dealing with the internal representation of time.

The macro US can be used by other functions for easy casting to US
allowing SensorUS to be written trivially.