Redesigned the ADC module to reflect design changes in RASWare
Incorporated various features such as pin definitions and timing functions
and redesigned the overall module to better match the standards that have
been created for RASWare.

Also adds background and busy wait functions that are more cpu expensive
than the contious reads but allow the option. The ADC module also assigns
the created ADC pins to the available ADC blocks in a staggered way to
balance to load and hopefully optimize performance.

Has not yet been tested for functionality.