Reassigned timers used by pwm
Because pwm is now setup to use all the wide timers, the timers used in time.c were remapped. It now uses Timer 5 as a full 32 bit one shot and the SysTick for clock updates. Unfortunately the SysTick on the LM4F is only 24bits, and with a speed of 50MHz means that it can't last for a full second. SysTick now should just interrupt as slow as possible and adjust the clock accordingly (which occurs about 3 times a second). None of these changes should be noticable outside of the time module.