Modified Makefile to NOT use the default thumb directory
Apparently for the Ubuntu/Debian versions of the arm library these just
switch to arm mode and bx to the arm versions of the float functions.

This is kiiiiiiiinda bad since the launchpad has no support for arm at
all and faults on the bx.

The best fix I can think of is to simply take to versions of the library
that work from arch and add another directory to check for those.
Currently I've put these on [toast](toast.projectgeky.com/rasware) as
CortexM4Libs so people can download them and use rasware on linux.

Will modify the readme later and may just add a bash script to do
everything for linux users.