Move detect-board script // Use the script in the makefile
RASLib is probably a better place for it rather than copying it into every project folder
(this is the same way the gdb-script is set up)

The makefile is set to invoke the script for the uart/run recipes rather than just set a
variable with the device path outside the recipes because the script is kind of slow on
macOS because of all the sed/awk/find/system_profiler calls (mainly system_profiler I think).

Rather than make every make invocation slower, we can just call the script when needed.

The script is a lil janky but it works and it's plenty fast on things that aren't macOS.